Set 'path' and reorder 'directory'
The contents of the option 'path' come in part from tweaking the output of
	cpp -v -xc++.

The contents of the option 'directory' reordered so that "~/tmp" is tried
first, ahead of ".", to avoid bothering git with .swp files in the working
directory. Of course that will be fixed in .gitignore as well.

modified:   .vimrc