Do inherited vim inits in the right order
Now .vimrc sources in .exrc first of all, and .gvimrc source in .vimrc first
of all too. Not that it matters much with their current contents, but it makes
sure that the inherited inits are done before we begin to mess around with
more specific ones.

modified:   .gvimrc
modified:   .vimrc