Ensure availability of qt code
For the compile time part, I just made sure that qt5-default and
qt5-qmake, along with qtchooser and qtcreator (the latter are optional,
and only for coding or debugging, not for just building the GUI).

For the run time, I took the ldd output on VerifySys, and traced back
each entry to a package.