cq/set-selective-display: tab is TAB-WIDTH spaces
The original version simply counted how many positions there were
between left margin and indentation margin--that is good only if the
indentation range is only spaces.

This version is more general than needed for
`cq/set-selective-display`: The arguments can be markers or integers;
the width may span lines, but it is always the horizontal offset.