Add git branch to shell mode line
I wanted to have a branch indication from any VC backend, but vc.el does
not generally consider a directory as belonging to a branch (yet). So
this works only for Git. Intentionally the mode line item starts with
"Git:" instead of "Git-", which I don't like because I use dashes in my
branch names.

Along the way commented out code in init-legacy.el that is not called
anymore.