Simplify and improve ruler
No more command line options; obtaining a footer is done by
piping to tac(1) in coreutils.

Additionally, document that a right-to-left ruler can be gotten
by piping to rev(1) in util-linux.

Finally, mark the first units position with a "1" instead of a ".",
to avoid having to remember if the ruler counts from 1 or from 0.
This is useful when reading bit layouts.