Re-enable the cedit setting
The official Ubuntu nvi still crashes when in the command editing window.
The nvi from the Ubuntu .tar.gz, when configured by

../dist/configure --enable-threads

does not have the problem, but complains about db3 create. That happens both
with the default configuration (above) and with --with-db=system.

So I will just leave the option activated, and maybe the function will work
correctly in the next release.