Updates to handle current gnome-terminal
I no longer install sakura as a workaround for having lost the ability to set
gnome-terminal's initial frame title. These changes try to make life
comfortable for those cases when gnome-terminal needs a title. They seem to
have forgotten that XTerm used to mean that you could have a GUI around a
terminal application, without compiling in to the app a GUI.

vv should now displace the last version of xo, which is due to move into the
Museum/.