New script vdiff
Takes up to four arguments, which it passes to vimdiff in the order given.

With a single filename, it passes to vimdiff the given filename with a "~"
appended, and then the given filename (good way to compare a file to its
backup).

Collects all flags and gives them to vimdiff up front. So,

$ vdiff vdiff

will vimdiff vdiff~ vdiff, if vdiff~ is exists and is readable, and complain if
there is no vdiff~.

On the other hand,

$ vdiff -g vdiff

with gvimdiff the same filenames.

The behavior with 3 or 4 arguments is that of vimdiff.