Disable AT_BRIDGE again; uncolorize terminal in vim
The NO_AT_BRIDGE=1 setting is also in /etc/environment now, in case I forget to
force it in my own .bash* files.

The VIM_TERMINAL env var is present and not empty when running under :terminal.
The color support there seems to be 16 out of 256 xterm-color256--the palette
doesn't work well with the dark background of the Solarized modes. Same used to
happen in earlier Emacsen, but the terminal emulators kept improving. So this
may be a temporary workaround only. (What, ten more years or so?)