First clone on Disco Dingo, no i386
As usual, the distro upgrade is a good time to remove obsolete packages. In this
case I took out an entire foreign architecture! I had added i386 in 2015 to run
some evals that were available only on 32-bit mode; the entire architecture has
not been needed for a long while.